refactor: further fixes and module changes
